feat(notifications): implement proper generation of notification emails for dashboard based notifications
