+++ backend/pkg/notification/pubkey_cache.go
@@ -43,7 +43,7 @@ func GetPubkeyForIndex(index uint64) ([]byte, error) {
 		if err != nil {
 			return nil, err
 		}
-		log.Infof("serving pubkey %x for validator %v from db", pubkey, index)
+		//log.Infof("serving pubkey %x for validator %v from db", pubkey, index)
 		return pubkey, nil
 	} else if err != nil {
 		return nil, err
